+++ assert/assertion.go
@@ -6,23 +6,43 @@ import (
 	"github.com/dogmatiq/testkit/render"
 )
 
-// OptionalAssertion is an interface that accepts all Assertion types, as well as
-// the Nothing value.
-type OptionalAssertion interface {
+// Operation is an enumeration of the test operations that can make assertions.
+type Operation int
+
+const (
+	// ExecuteCommandOperation is an operation that makes assertions about what
+	// happens when a command is executed.
+	ExecuteCommandOperation Operation = iota
+
+	// RecordEventOperation is an operation that makes assertions about what
+	// happens when an event is recorded.
+	RecordEventOperation
+
+	// AdvanceTimeOperation is an operation that makes assertions about what
+	// happens when the engine time advances.
+	AdvanceTimeOperation
+
+	// CallOperation is an operation that makes assertions about what happens
+	// when a user-defined function is invoked.
+	CallOperation
+)
+
+// An Assertion is a predicate for determining whether some specific criteria
+// was met during a test.
+type Assertion interface {
 	fact.Observer
 
 	// Begin is called to prepare the assertion for a new test.
 	//
-	// c is the comparator used to compare messages and other entities.
-	Begin(c compare.Comparator)
+	// op is the operation that is making the assertion. c is the comparator
+	// used to compare messages and other entities.
+	Begin(op Operation, c compare.Comparator)
 
 	// End is called once the test is complete.
 	End()
 
-	// TryOk returns true if the assertion passed.
-	//
-	// If asserted is false, the assertion was a no-op and ok is meaningless.
-	TryOk() (ok bool, asserted bool)
+	// Ok returns true if the assertion passed.
+	Ok() bool
 
 	// BuildReport generates a report about the assertion.
 	//
@@ -32,25 +52,19 @@ type OptionalAssertion interface {
 	BuildReport(ok bool, r render.Renderer) *Report
 }
 
-// An Assertion is a predicate for determining whether some specific criteria
-// was met during a test.
-type Assertion interface {
-	OptionalAssertion
+// Nothing is an assertion that has no requirements.
+var Nothing Assertion = nothingAssertion{}
 
-	// Ok returns true if the assertion passed.
-	Ok() bool
-}
+type nothingAssertion struct{}
 
-// Nothing is an "optional assertion" that always passes and does not build a
-// report.
-var Nothing OptionalAssertion = noopAssertion{}
-
-type noopAssertion struct{}
-
-func (noopAssertion) Notify(fact.Fact)         {}
-func (noopAssertion) Begin(compare.Comparator) {}
-func (noopAssertion) End()                     {}
-func (noopAssertion) TryOk() (bool, bool)      { return false, false }
-func (noopAssertion) BuildReport(bool, render.Renderer) *Report {
-	panic("not implemented")
+func (nothingAssertion) Notify(fact.Fact)                    {}
+func (nothingAssertion) Begin(Operation, compare.Comparator) {}
+func (nothingAssertion) End()                                {}
+func (nothingAssertion) Ok() bool                            { return true }
+func (nothingAssertion) BuildReport(ok bool, _ render.Renderer) *Report {
+	return &Report{
+		TreeOk:   ok,
+		Ok:       true,
+		Criteria: "no requirement",
+	}
 }